==> 03_round_trips.Rmd <==
---
title: "Round Trips"
author: "Doug Antibus"
date: "2023-12-28"
output: 
  html_document:
    toc: true
    number_sections: false
    theme:
      version: 5
      bootswatch: pulse
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(message = FALSE)
knitr::opts_chunk$set(warning = FALSE)
source("00_libraries.r")
```

### Identifying round trips

Another feature we might be interested in is whether or not any given trip represents
a round trip.  For example, members and casuals might differ in their tendency to 
take round trips.

Start by reading in the most recent data frame `bikes_1`.

```{r, cache=TRUE}
bikes <- read_csv(here("data", "bikes_1.csv")) 
```

Identify records representing round trips, then `mutate` creating a 
new Boolean variable `round_trip`. 

```{r}
rt <- bikes %>% filter(bikes$start_name == bikes$end_name) %>% 
  select(ride_id, start_name, end_name, member_casual)

rt_1 <- rt %>%  mutate(round_trip = "yes") %>% select(ride_id, round_trip)
glimpse(rt_1)
```

Join round trip data to the main data frame using `ride_id`.  If the round 
trip field is NA (i.e. not a round trip), set the variable to be "no", else set 
it to be "yes".  We put this into a new dataframe called `bikes_new`.

```{r}
bikes_new <- bikes %>% 
  left_join(rt_1, by = "ride_id") %>% 
  mutate(round_trip = ifelse(is.na(round_trip), "no", "yes"))
glimpse(bikes_new)
```

Finally, write bikes_new to a file, `bikes_2.csv`.

```{r, eval = FALSE, cache=TRUE}
bikes_new %>% write_csv(here("data", "bikes_2.csv"))
```
